+++ CHANGES.md
@@ -4,6 +4,7 @@
 - The DatePicker mode now adds a base style called "calendar-date-picker-input" to the input box.
 - Added new setting "datePickerSelectedDateFormat", which allows the selected date (in the input field) in DatePicker mode to be formatted (see documentation).
 - Fixed a fault that prevented the public function setSelectedDatePickerDate() from setting the DatePicker date.
+- The "Search" button on all toolbars is now only visible when events are available.
 - 
 
 ### Version 1.5.3: